requirements.txt
django
channels
daphne
channels-redis
django-cors-headers
redis
psycopg2-binary
psycopg
djangorestframework-simplejwt